requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
noise>=1.2.2
pdfplumber>=0.10.0
python-docx>=1.1.0
img2pdf>=0.4.4
gradio>=4.0.0
kaggle>=1.6.0
tqdm>=4.66.0
opencv-python>=4.8.0
lxml>=4.9.0
